+++ src/main/resources/root-context.xml
@@ -26,5 +26,5 @@
 		<constructor-arg index="0" ref="sqlSession"/>
 	</bean>
 	
-	<util:properties id="oauthInfo" location="classpath:/config/app.properties" />
+	<util:properties id="keys" location="classpath:/config/app.properties" />
 </beans>
