+++ src/main/webapp/resources/css/common/main.css
@@ -209,10 +209,26 @@ hr{
         font-size: 0.8rem;
         padding: 0.2rem 0;
     }
-
+	
+	.img-frame {
+		width: 170px;
+        height: 175px;
+        object-fit: cover;
+        overflow:hidden;
+	}
+    .list-item img{
+    	transition: transform 1s;
+        overflow:hidden;
+    }
+    
+    .list-item:hover img{
+    	transform: scale(1.3);
+        transition: transform 1s;
+    }
+    
     .list-thumb {
         width: 170px;
-        height: 175px;
+        height: 100%;
         object-fit: cover;
     }
